requirements.txt
# Required Python packages for Pizzini Social Media Automation

# Social Media APIs
tweepy>=4.14.0              # Twitter/X API client
instagrapi>=1.19.0          # Instagram API client

# Image processing for Instagram posts
Pillow>=10.0.0              # Image generation and manipulation

# Scheduling
schedule>=1.2.0             # Task scheduling

# Security and encryption
cryptography>=41.0.0        # Credential encryption

# HTTP requests
requests>=2.31.0            # HTTP library for various API calls

# Optional: Additional utilities
python-dotenv>=1.0.0        # Environment variable management (optional)